fix: Convert Long to Double for formatCurrencyCompact function
Fix type mismatch where todayRevenue (Long) was passed to
formatCurrencyCompact which expects Double.

+++ ccb-android/app/src/main/java/com/techres/ccb/presentation/screens/dashboard/DashboardScreen.kt
@@ -656,7 +656,7 @@ private fun MobileDashboardContent(
                             )
                             Spacer(modifier = Modifier.width(4.dp))
                             Text(
-                                text = formatCurrencyCompact(uiState.todayRevenue),
+                                text = formatCurrencyCompact(uiState.todayRevenue.toDouble()),
                                 color = Color(0xFF4CAF50),
                                 fontWeight = FontWeight.Bold,
                                 fontSize = 14.sp
